requirements_poly.txt
PyGithub>=1.55
numpy>=1.21.0
requests>=2.25.1